test(motion): recategorize and clean up browser tests
